chore: extract common eslint configuration
We still have files for package-specific customization if & when we want
it.
