+++ package.json
@@ -16,7 +16,14 @@
     "release": "pnpm -r release && release-it"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.29.0",
     "@release-it-plugins/workspaces": "^4.2.0",
+    "@typescript-eslint/eslint-plugin": "^7.0.0",
+    "@typescript-eslint/parser": "^7.0.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "globals": "^14.0.0",
     "npm-run-all": "^4.1.5",
     "release-it": "^17.0.0",
     "sort-package-json": "^3.2.1",
