+++ packages/local-mcp-server/eslint.config.mjs
@@ -1,101 +1,3 @@
-import typescriptEslint from '@typescript-eslint/eslint-plugin';
-import prettier from 'eslint-plugin-prettier';
-import globals from 'globals';
-import tsParser from '@typescript-eslint/parser';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import js from '@eslint/js';
-import { FlatCompat } from '@eslint/eslintrc';
+import { commonConfig } from '../../eslint.common.mjs';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-  baseDirectory: __dirname,
-  recommendedConfig: js.configs.recommended,
-  allConfig: js.configs.all,
-});
-
-export default [
-  {
-    ignores: [
-      '**/.eslintrc.js',
-      '**/prettier.config.js',
-      '**/build',
-      '**/node_modules',
-      '**/package-log.json',
-    ],
-  },
-  ...compat.extends(
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended',
-  ),
-  {
-    plugins: {
-      '@typescript-eslint': typescriptEslint,
-      prettier,
-    },
-
-    languageOptions: {
-      globals: {
-        ...globals.node,
-        ...globals.mocha,
-      },
-
-      parser: tsParser,
-      ecmaVersion: 2019,
-      sourceType: 'module',
-
-      parserOptions: {
-        project: './tsconfig.json',
-      },
-    },
-
-    rules: {
-      '@typescript-eslint/ban-ts-ignore': 'off',
-      '@typescript-eslint/type-annotation-spacing': 'off',
-      '@typescript-eslint/explicit-function-return-type': 'off',
-      'no-unused-vars': 'off',
-
-      '@typescript-eslint/no-unused-vars': [
-        'warn',
-        {
-          args: 'none',
-        },
-      ],
-
-      '@typescript-eslint/adjacent-overload-signatures': 'error',
-      camelcase: 'off',
-      '@typescript-eslint/consistent-type-assertions': 'error',
-      'no-array-constructor': 'off',
-      '@typescript-eslint/no-array-constructor': 'error',
-      'no-empty': 'off',
-      'no-empty-function': 'off',
-      '@typescript-eslint/explicit-module-boundary-types': 'off',
-      '@typescript-eslint/no-empty-function': 'off',
-      '@typescript-eslint/no-empty-interface': 'error',
-      '@typescript-eslint/no-explicit-any': 'off',
-      '@typescript-eslint/no-inferrable-types': 'error',
-      '@typescript-eslint/no-misused-new': 'error',
-      '@typescript-eslint/no-namespace': 'error',
-      '@typescript-eslint/no-non-null-assertion': 'off',
-      '@typescript-eslint/no-this-alias': 'error',
-      'no-use-before-define': 'off',
-      '@typescript-eslint/no-use-before-define': 'off',
-      '@typescript-eslint/no-var-requires': 'error',
-      '@typescript-eslint/prefer-namespace-keyword': 'error',
-      '@typescript-eslint/triple-slash-reference': 'error',
-      '@typescript-eslint/no-floating-promises': ['error'],
-      'no-var': 'error',
-      'prefer-const': 'error',
-      'prefer-rest-params': 'error',
-      'prefer-spread': 'error',
-    },
-  },
-  {
-    files: ['**/*.js'],
-
-    rules: {
-      '@typescript-eslint/no-var-requires': 'off',
-    },
-  },
-];
+export default commonConfig;
